@@ -43,26 +43,26 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular-devkit/build-ng-packagr": "~0.7.0",
-    "@angular/animations": "^7.2.16",
-    "@angular/cli": "~7.3.10",
-    "@angular/common": "^7.2.16",
-    "@angular/compiler": "^7.2.16",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/core": "^7.2.16",
-    "@angular/forms": "^7.2.16",
-    "@angular/http": "^7.2.16",
-    "@angular/language-service": "^7.2.16",
-    "@angular/platform-browser": "^7.2.16",
-    "@angular/platform-browser-dynamic": "^7.2.16",
-    "@angular/router": "^7.2.16",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular-devkit/build-ng-packagr": "~0.803.29",
+    "@angular/animations": "^8.2.14",
+    "@angular/cli": "~8.3.29",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@types/extend": "^3.0.0",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
     "@types/prismjs": "~1.4.18",
     "bowser": "^1.7.2",
+    "codelyzer": "^5.0.1",
     "core-js": "^2.5.4",
     "date-fns": "2.0.0-alpha.1",
     "element-closest": "^2.0.2",
@@ -79,16 +79,16 @@
     "karma-min-reporter": "^0.1.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-remap-istanbul": "~0.4.0",
-    "ng-packagr": "^3.0.0",
+    "ng-packagr": "^5.4.0",
     "popper.js": "^1.14.0",
     "prismjs": "^1.29.0",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.6.7",
     "ts-node": "~5.0.1",
-    "tsickle": ">=0.29.0",
+    "tsickle": "^0.37.0",
     "tslib": "^1.9.0",
     "tslint": "~5.9.1",
-    "typescript": "~3.2.4",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.26"
+    "typescript": "~3.5.3",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
   }
 }
